Project: Latesummer Description:增删改查示例完成

+++ src/main/java/com/latesummer/domain/LearnResouce.java
@@ -14,22 +14,22 @@ import com.latesummer.common.domain.BaseModel;
  */
 @Entity
 @Table(name = "learn_resource")
-public class LearnResouce  implements BaseModel<Integer>{
+public class LearnResouce implements BaseModel<Long>{
 
 	private static final long serialVersionUID = -91221104520172449L;
 	
 	@Id
 	@GeneratedValue
-    private Integer id;
+    private Long id;
 	private String author;
     private String title;
     private String url;
     
-    public Integer getId() {
+    public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
