+++ src/main/java/com/latesummer/web/controller/UserController.java
@@ -27,7 +27,7 @@ public class UserController {
     static Map<Long, User> users = Collections.synchronizedMap(new HashMap<Long, User>()); 
 
 	@RequestMapping(value = "/add/{id}/{name}/{age}")
-	public User addUser(@PathVariable Integer id, @PathVariable String name, @PathVariable Integer age) {
+	public User addUser(@PathVariable Long id, @PathVariable String name, @PathVariable Integer age) {
 		User user = new User();
 		user.setId(id);
 		user.setName(name);
